+++ Tests/Public/Remove-TeamViewerManager.Tests.ps1
@@ -1,7 +1,7 @@
 BeforeAll {
-    . "$PSScriptRoot/../../TeamViewerPS/Public/Remove-TeamViewerManager.ps1"
+    . "$PSScriptRoot/../../docs/Cmdlets/Public/Remove-TeamViewerManager.ps1"
 
-    @(Get-ChildItem -Path "$PSScriptRoot/../../TeamViewerPS/Private/*.ps1") | `
+    @(Get-ChildItem -Path "$PSScriptRoot/../../docs/Cmdlets/Private/*.ps1") | `
         ForEach-Object { . $_.FullName }
 
     $testApiToken = [securestring]@{}