+++ Tests/Public/Remove-TeamViewerPolicyFromManagedDevice.Tests.ps1
@@ -1,8 +1,8 @@
 BeforeAll {
-    . "$PSScriptRoot/../../TeamViewerPS/Public/Remove-TeamViewerPolicyFromManagedDevice.ps1"
-    . "$PSScriptRoot/../../TeamViewerPS/TeamViewerPS.Types.ps1"
+    . "$PSScriptRoot/../../docs/Cmdlets/Public/Remove-TeamViewerPolicyFromManagedDevice.ps1"
+    . "$PSScriptRoot/../../docs/TeamViewerPS.Types.ps1"
 
-    @(Get-ChildItem -Path "$PSScriptRoot/../../TeamViewerPS/Private/*.ps1") | `
+    @(Get-ChildItem -Path "$PSScriptRoot/../../docs/Cmdlets/Private/*.ps1") | `
         ForEach-Object { . $_.FullName }
 
     $testApiToken = [securestring]@{}
@@ -52,4 +52,3 @@ Describe 'Remove-TeamViewerPolicyFromManagedDevice' {
                 -Device $testDeviceId -PolicyType 2 } | Should -Throw
     }
 }
-